improved AStarSolver, less memory, slightly faster
AStarNode doesn't have two long[] arrays anymore, instead it contains two index values into a compact storage area.
This reduces the memory overhead of the array objects: instead of using many small arrays, there are now a few large arrays that store the A* search states.

Furthermore, the hash table that's used to prune duplicate search paths, now uses the same storage area as well.
This reduces memory usage by not storing the state twice anymore, in the priority queue and in the hash table.
For a speedup of the operations key-lookup and table-resize, the hash table stores the storage index and the hash value.